+++ src/effects/effectTypes/finalImageEffects/pixelate/invoke.js
@@ -1,25 +1,28 @@
 import {getWorkingDirectory} from "../../../../core/GlobalSettings.js";
-import {randomId} from "../../../../core/math/random.js";
+import {getRandomIntInclusive, randomId} from "../../../../core/math/random.js";
 import Jimp from "jimp";
 import {findValue} from "../../../../core/math/findValue.js";
 import fs from "fs";
 
 export const pixelate = async (layer, data, currentFrame, totalFrames) => {
-    const filename = getWorkingDirectory() + 'pixelate' + randomId() + '.png';
+    const theGlitch = getRandomIntInclusive(0, 100);
+    if (theGlitch < data.glitchChance) {
+        const filename = getWorkingDirectory() + 'pixelate' + randomId() + '.png';
 
-    await layer.toFile(filename);
+        await layer.toFile(filename);
 
-    const jimpImage = await Jimp.read(filename);
+        const jimpImage = await Jimp.read(filename);
 
-    const pixelateGaston = Math.floor(findValue(data.lower, data.upper, data.times, totalFrames, currentFrame));
+        const pixelateGaston = Math.floor(findValue(data.lower, data.upper, data.times, totalFrames, currentFrame));
 
-    if (pixelateGaston > 0) {
-        await jimpImage.pixelate(pixelateGaston);
+        if (pixelateGaston > 0) {
+            await jimpImage.pixelate(pixelateGaston);
+        }
+
+        await jimpImage.writeAsync(filename);
+
+        await layer.fromFile(filename);
+
+        fs.unlinkSync(filename);
     }
-
-    await jimpImage.writeAsync(filename);
-
-    await layer.fromFile(filename);
-
-    fs.unlinkSync(filename);
 }
